sessions/session-03-rag/participants/nithvit/requirements.txt
google-generativeai>=0.8.5
google-genai>=1.51.0
chromadb>=1.3.4
langchain>=1.0.7
langchain-text-splitters>=1.0.0
streamlit>=1.51.0
python-dotenv>=1.2.1
python-docx>=0.8.11
